requirements.txt
flask==3.0.0
google-generativeai==0.8.3
python-dotenv==1.0.0
requests==2.31.0
duckduckgo-search==6.3.5

